added reporting of system time

+++ manager_3/io/fileWriter.c
@@ -2,9 +2,9 @@
 #include <stdint.h>
 #include <stdio.h>
 
-void writeToFileSingle(FILE* fin, pthread_mutex_t* mutex, uint8_t message){
+void writeToFileSingle(FILE* fin, pthread_mutex_t* mutex, long message){
     pthread_mutex_lock(mutex);
-    fprintf(fin,"%hhu\n", message);
+    fprintf(fin,"%ld\n", message);
     fflush(fin);
     pthread_mutex_unlock(mutex);
 }
