+++ manager_3/io/fileWriter.h
@@ -2,5 +2,5 @@
 #include <stdint.h>
 #include <stdio.h>
 
-void writeToFileSingle(FILE* fin, pthread_mutex_t* mutex, uint8_t message);
+void writeToFileSingle(FILE* fin, pthread_mutex_t* mutex, long message);
 void writeToFileMultiple(FILE* fin, pthread_mutex_t* mutex, int size, uint8_t *message);
